requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-jose
cryptography
passlib
bcrypt
python-dotenv
Pillow
qrcode
pydantic
email-validator